// src/api/gen/types/FieldView.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type FieldView = {
    code?: string;
    component?: string;
    fieldType?: string;
    label?: string;
    options?: string[];
    required?: boolean;
    requiresModule?: string;
    validation?: {
        [key: string]: unknown;
    };
};